refactor: ProfileSettingsModal을 컴파운드 패턴으로 마이그레이션
